[TASK] Plant ID verification at page start #2
Implemented consistency check at the start of page loading.

Consistency check tests data for the following rules:
1. For each plant its `id` attribute should be equal
   to the dictionary key under which it resides.
2. For each mutation parent of a plant its `id` should
   be able to be used as a key of the plant dictionary.
   Exception to rule 2 is id = "any" which is not a
   reference to a specific plant.

The check is done for every plant and every mutation parent.
When check encounters an error it does not stop and continues
untill all plants are checked.
The result of a check is a list of errors which normally
should be empty if no errors are found.

If any error is spotted they all are printed on the page header.